backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.0.5
Deprecated==1.2.13
distro==1.8.0
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
environs==9.5.0
exceptiongroup==1.0.0
factory-boy==3.2.1
Faker==12.0.1
freezegun==1.2.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
kombu==5.2.4
marshmallow==3.19.0
mixer==7.2.2
netaddr==0.8.0
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-docker==1.0.1
pytest-factoryboy==2.5.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.5
redis==4.3.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.21.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yookassa==2.3.2
